+++ src/collective/feedback/testing.py
@@ -1,55 +1,52 @@
-# -*- coding: utf-8 -*-
-from plone.app.robotframework.testing import REMOTE_LIBRARY_BUNDLE_FIXTURE
+from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_FIXTURE
 from plone.app.testing import applyProfile
 from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
-from plone.app.testing import PLONE_FIXTURE
 from plone.app.testing import PloneSandboxLayer
-from plone.testing import z2
+from plone.testing.zope import WSGI_SERVER_FIXTURE
 
 import collective.feedback
-import plone.app.dexterity
+import collective.honeypot
+import collective.honeypot.config
 import plone.restapi
 import souper.plone
 
 
-class CollectiveFeedbackLayer(PloneSandboxLayer):
-    defaultBases = (PLONE_FIXTURE,)
+collective.honeypot.config.EXTRA_PROTECTED_ACTIONS = set(["feedback-add"])
+collective.honeypot.config.HONEYPOT_FIELD = "honey"
+
+
+class TestLayer(PloneSandboxLayer):
+    defaultBases = (PLONE_APP_CONTENTTYPES_FIXTURE,)
 
     def setUpZope(self, app, configurationContext):
         # Load any other ZCML that is required for your tests.
         # The z3c.autoinclude feature is disabled in the Plone fixture base
         # layer.
-
-        self.loadZCML(package=plone.app.dexterity)
         self.loadZCML(package=plone.restapi)
-        self.loadZCML(package=collective.feedback)
+        self.loadZCML(package=collective.honeypot)
         self.loadZCML(package=souper.plone)
+        self.loadZCML(package=collective.feedback)
 
     def setUpPloneSite(self, portal):
         applyProfile(portal, "collective.feedback:default")
 
 
-COLLECTIVE_FEEDBACK_FIXTURE = CollectiveFeedbackLayer()
+FIXTURE = TestLayer()
 
 
-COLLECTIVE_FEEDBACK_INTEGRATION_TESTING = IntegrationTesting(
-    bases=(COLLECTIVE_FEEDBACK_FIXTURE,),
+INTEGRATION_TESTING = IntegrationTesting(
+    bases=(FIXTURE,),
     name="CollectiveFeedbackLayer:IntegrationTesting",
 )
 
 
-COLLECTIVE_FEEDBACK_FUNCTIONAL_TESTING = FunctionalTesting(
-    bases=(COLLECTIVE_FEEDBACK_FIXTURE,),
+FUNCTIONAL_TESTING = FunctionalTesting(
+    bases=(FIXTURE,),
     name="CollectiveFeedbackLayer:FunctionalTesting",
 )
 
-
-COLLECTIVE_FEEDBACK_ACCEPTANCE_TESTING = FunctionalTesting(
-    bases=(
-        COLLECTIVE_FEEDBACK_FIXTURE,
-        REMOTE_LIBRARY_BUNDLE_FIXTURE,
-        z2.ZSERVER_FIXTURE,
-    ),
-    name="CollectiveFeedbackLayer:AcceptanceTesting",
+RESTAPI_TESTING = FunctionalTesting(
+    bases=(FIXTURE, WSGI_SERVER_FIXTURE),
+    name="CollectiveFeedbackLayer:RestAPITesting",
 )